task/347567: The code of the validation process was optimized, and a trait for the form was created.
